+++ sqlstore/src/main/resources/db/migration/V8000__kitlot_expand.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Library MODIFY COLUMN kitLot varchar(255);