+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/TransferService.java
@@ -10,6 +10,7 @@ import com.eaglegenomics.simlims.core.User;
 import uk.ac.bbsrc.tgac.miso.core.data.Lab;
 import uk.ac.bbsrc.tgac.miso.core.data.Project;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.transfer.Transfer;
+import uk.ac.bbsrc.tgac.miso.core.data.impl.transfer.TransferLibrary;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.transfer.TransferSample;
 
 public interface TransferService extends DeleterService<Transfer>, ListService<Transfer>, SaveService<Transfer> {
@@ -19,5 +20,7 @@ public interface TransferService extends DeleterService<Transfer>, ListService<T
   public List<Transfer> listByProperties(Lab sender, Group recipient, Project project, Date transferTime) throws IOException;
 
   public void addTransferSample(TransferSample transferSample) throws IOException;
+
+  public void addTransferLibrary(TransferLibrary transferLibrary) throws IOException;
 
 }