+++ sqlstore/src/main/resources/db/migration/V1260__auto_main.sql
@@ -1,3 +1,7 @@
+-- kitlot_expand
+ALTER TABLE Library MODIFY COLUMN kitLot varchar(255);
+
+-- transfer_volume
 ALTER TABLE Transfer_Sample
   ADD COLUMN distributedVolume DECIMAL(16,10),
   ADD COLUMN distributedBoxAlias varchar(255),
@@ -17,3 +21,4 @@ ALTER TABLE Transfer_Pool
   ADD COLUMN distributedVolume DECIMAL(16,10),
   ADD COLUMN distributedBoxAlias varchar(255),
   ADD COLUMN distributedBoxPosition varchar(3);
+