+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 Changes:
 
-* Extend 'Kit Lot' field on Libraries to 1000 characters
+* Extend 'Kit Lot' field on Libraries to 255 characters
 * Add front-end error checking for Kit Lot length
 
 # 1.25.0