+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@ Changes:
 
 * Extend 'Kit Lot' field on Libraries to 255 characters
 * Add front-end error checking for Kit Lot length
+* Improved error message when attempting to delete a library aliquot that
+  is included in a pool order
 
 # 1.25.0