+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ Changes:
 * Add front-end error checking for Kit Lot length
 * Improved error message when attempting to delete a library aliquot that
   is included in a pool order
+* Improved performance of library receipt
 
 # 1.25.0